Optimize taurize proc with same part and color

+++ code/modules/surgery/bodyparts/helpers.dm
@@ -272,13 +272,18 @@
 	set_resting(FALSE)
 
 /mob/living/carbon/proc/Taurize(taur_type = /obj/item/bodypart/taur/horse, color = "#ffffff")
+	// Same taur part short circuit to save on taurize cost because it occupies up to 8 - 9 ms out of a 20 ms call of preview
+	var/obj/item/bodypart/taur/existing = get_taur_tail()
+	if(existing && existing.type == taur_type && existing.taur_color == color)
+		return
+
 	for(var/X in bodyparts)
 		var/obj/item/bodypart/O = X
 		// drop taur tails too
 		if(O.body_part == LEG_LEFT || O.body_part == LEG_RIGHT || O.body_zone == BODY_ZONE_TAUR)
 			O.drop_limb(1)
 			qdel(O)
-	
+
 	var/obj/item/bodypart/taur/T = new taur_type()
 	T.taur_color = color
 	T.attach_limb(src)
